refactor: rename aci.ts to Contract.ts
As preparation for future refactoring.
