+++ src/account/Generalized.ts
@@ -23,7 +23,7 @@ import {
 } from '../utils/errors';
 import { decode, Encoded } from '../utils/encoder';
 import { getAccount } from '../chain';
-import getContractInstance from '../contract/aci';
+import getContractInstance from '../contract/Contract';
 import { _buildTx } from '../tx';
 import { Tag } from '../tx/builder/constants';
 import { buildTx } from '../tx/builder';
