+++ src/index.ts
@@ -29,7 +29,7 @@ export * from './utils/amount-formatter';
 export * from './utils/hd-wallet';
 export { encode, decode, Encoding } from './utils/encoder';
 export * from './aens';
-export { default as getContractInstance } from './contract/aci';
+export { default as getContractInstance } from './contract/Contract';
 export * from './oracle';
 export * from './spend';
 export * from './contract/ga';
